New upstream version 0.9.12.3

+++ src/conf/domain_nwfilter.c
@@ -14,8 +14,8 @@
  * Lesser General Public License for more details.
  *
  * You should have received a copy of the GNU Lesser General Public
- * License along with this library.  If not, see
- * <http://www.gnu.org/licenses/>.
+ * License along with this library; if not, write to the Free Software
+ * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307  USA
  *
  * Author: Stefan Berger <stefanb@us.ibm.com>
  */
@@ -53,7 +53,7 @@ virDomainConfNWFilterTeardown(virDomainNetDefPtr net) {
 
 void
 virDomainConfVMNWFilterTeardown(virDomainObjPtr vm) {
-    size_t i;
+    int i;
 
     if (nwfilterDriver != NULL) {
         for (i = 0; i < vm->def->nnets; i++)
